requirements.txt
asgiref==3.9.1
attrs==25.3.0
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.6
django-eose==0.3.1
django-redis==6.0.0
django-summernote==0.8.20.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
mysql-connector-python==9.4.0
mysqlclient==2.2.7
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
psutil==7.0.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
